+++ code/src/ui/graphic/table/action/registry-row-action.tsx
@@ -179,6 +179,7 @@ export default function RegistryRowAction(
         className="ml-2"
         isOpen={isActionMenuOpen}
         setIsOpen={setIsActionMenuOpen}
+        aria-label={`${dict.title.actions}, ${props.row.id}`}
       >
         <div className="flex flex-col space-y-8 lg:space-y-4 ">
           {isSubmissionOrGeneralPage && (