+++ CHANGELOG.md
@@ -2,6 +2,15 @@
 
 [//]: # 'Note that version headers need to start with "## " characters to be picked up by some automated scripts'
 
+## 5.69.2
+
+### Changes
+
+- Accessiblity improvements: 
+- Standardized the date display format
+- Aria-labels on regsitry row actions, bulk approvement
+- Made the table pagination as a `nav` HTML element
+
 ## 5.69.1
 
 ### Changes